requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary
mcp>=1.0.0
openai>=1.3.5
groq>=0.4.1